+++ tests/exactly_once_subscription.rs
@@ -50,13 +50,7 @@ fn main() -> std::io::Result<()> {
 
         if !published && mqtt.client().can_publish(QoS::ExactlyOnce) {
             mqtt.client()
-                .publish(
-                    Publication::new("Ping")
-                        .topic("data")
-                        .qos(QoS::ExactlyOnce)
-                        .finish()
-                        .unwrap(),
-                )
+                .publish(Publication::new("data", b"Ping").qos(QoS::ExactlyOnce))
                 .unwrap();
             log::info!("Publishing message");
             published = true;